Create Commuting
Without deep knowledge of the variables, I have created this profile following the instructions in https://github.com/poutnikl/Brouter-profiles/wiki/Determining-optimal-bicycle-profile Cooked. 

I hope this might be useful to somebody and would be glad if someone could revise it and include it together with the other profiles. 

I still have doubts about the avoid_unsafe variable, which I decided not to touch.